LayoutTests/fast/forms/label/labels-remove-htmlFor-attribute-expected.txt

Test that the labels list of a form control is dynamically updated when the htmlFor attribute of a label is removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS labels.length is 1
PASS labels.length is 0
PASS successfullyParsed is true

TEST COMPLETE
